+++ flutter_inappwebview_platform_interface/lib/src/platform_proxy_controller.g.dart
@@ -6,11 +6,64 @@ part of 'platform_proxy_controller.dart';
 // ExchangeableObjectGenerator
 // **************************************************************************
 
-///Class that represents the settings used to configure the [PlatformProxyController].
-///
-///**Officially Supported Platforms/Implementations**:
-///- Android native WebView ([Official API - ProxyConfig](https://developer.android.com/reference/androidx/webkit/ProxyConfig))
-class ProxySettings {
+class IOSProxySettings {
+  /// A Boolean that indicates whether or not a proxy configuration allows failover to non-proxied connections.
+  /// Failover isn’t allowed by default.
+  bool allowFailover;
+  List<String> excludedDomains;
+  List<String> matchDomains;
+
+  ///Proxy is a string in the format `[scheme://]host[:port]`.
+  ///Scheme is optional, if present must be `HTTP`, `HTTPS` or [SOCKS](https://tools.ietf.org/html/rfc1928) and defaults to `HTTP`.
+  String proxyUrl;
+  IOSProxySettings(
+      {this.allowFailover = false,
+      this.excludedDomains = const [],
+      this.matchDomains = const [],
+      this.proxyUrl = ''});
+
+  ///Gets a possible [IOSProxySettings] instance from a [Map] value.
+  static IOSProxySettings? fromMap(Map<String, dynamic>? map) {
+    if (map == null) {
+      return null;
+    }
+    final instance = IOSProxySettings();
+    instance.allowFailover = map['allowFailover'];
+    instance.excludedDomains =
+        List<String>.from(map['excludedDomains']!.cast<String>());
+    instance.matchDomains =
+        List<String>.from(map['matchDomains']!.cast<String>());
+    instance.proxyUrl = map['proxyUrl'];
+    return instance;
+  }
+
+  ///Converts instance to a map.
+  Map<String, dynamic> toMap() {
+    return {
+      "allowFailover": allowFailover,
+      "excludedDomains": excludedDomains,
+      "matchDomains": matchDomains,
+      "proxyUrl": proxyUrl,
+    };
+  }
+
+  ///Converts instance to a map.
+  Map<String, dynamic> toJson() {
+    return toMap();
+  }
+
+  ///Returns a copy of IOSProxySettings.
+  IOSProxySettings copy() {
+    return IOSProxySettings.fromMap(toMap()) ?? IOSProxySettings();
+  }
+
+  @override
+  String toString() {
+    return 'IOSProxySettings{allowFailover: $allowFailover, excludedDomains: $excludedDomains, matchDomains: $matchDomains, proxyUrl: $proxyUrl}';
+  }
+}
+
+class AndroidProxySettings {
   ///List of bypass rules.
   ///
   ///A bypass rule describes URLs that should skip proxy override settings and make a direct connection instead. These can be URLs or IP addresses. Wildcards are accepted.
@@ -59,7 +112,7 @@ class ProxySettings {
   ///
   ///**NOTE**: available only if [WebViewFeature.PROXY_OVERRIDE_REVERSE_BYPASS] feature is supported.
   bool reverseBypassEnabled;
-  ProxySettings(
+  AndroidProxySettings(
       {this.bypassRules = const [],
       this.bypassSimpleHostnames,
       this.directs = const [],
@@ -67,12 +120,12 @@ class ProxySettings {
       this.removeImplicitRules,
       this.reverseBypassEnabled = false});
 
-  ///Gets a possible [ProxySettings] instance from a [Map] value.
-  static ProxySettings? fromMap(Map<String, dynamic>? map) {
+  ///Gets a possible [AndroidProxySettings] instance from a [Map] value.
+  static AndroidProxySettings? fromMap(Map<String, dynamic>? map) {
     if (map == null) {
       return null;
     }
-    final instance = ProxySettings(
+    final instance = AndroidProxySettings(
       bypassSimpleHostnames: map['bypassSimpleHostnames'],
       removeImplicitRules: map['removeImplicitRules'],
     );
@@ -102,13 +155,13 @@ class ProxySettings {
     return toMap();
   }
 
-  ///Returns a copy of ProxySettings.
-  ProxySettings copy() {
-    return ProxySettings.fromMap(toMap()) ?? ProxySettings();
+  ///Returns a copy of AndroidProxySettings.
+  AndroidProxySettings copy() {
+    return AndroidProxySettings.fromMap(toMap()) ?? AndroidProxySettings();
   }
 
   @override
   String toString() {
-    return 'ProxySettings{bypassRules: $bypassRules, bypassSimpleHostnames: $bypassSimpleHostnames, directs: $directs, proxyRules: $proxyRules, removeImplicitRules: $removeImplicitRules, reverseBypassEnabled: $reverseBypassEnabled}';
+    return 'AndroidProxySettings{bypassRules: $bypassRules, bypassSimpleHostnames: $bypassSimpleHostnames, directs: $directs, proxyRules: $proxyRules, removeImplicitRules: $removeImplicitRules, reverseBypassEnabled: $reverseBypassEnabled}';
   }
 }